archive/requirements.txt
asyncio
aiofiles
aiohttp-retry
asyncssh
pyarrow
tqdm
requests
omegaconf
pandas
python-dwca-reader
Pillow
scikit-learn
pyremotedata
